fsl-image-mfgtool-initramfs: Add image
Small image to be used with Manufacturing Tool (mfg-tool) in a
production environment.

Series-to: meta-fsl
Series-prefix: meta-fsl-arm
Cover-letter:
Freescale MfgTool artifacts support
We've been cooking this patchset for a while at O.S. Systems and I
think it is ready for pristine time.

This allow us to generate the initramfs, kernel and u-boot artifacts
for use with Freescale MfgTool. This has been deployed for some of our
customers already and it's working fine for them and this allowed us
to improve and make it generic as possible for use in
customer-specific layers.

Please review and test it.
END

Change-Id: Ibcab8826a70339d0cbf507311c6268ad74491df5
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>